+++ layout/explore/component.jsx
@@ -1,5 +1,6 @@
 import { useState, useMemo, useCallback } from 'react';
 import Head from 'next/head';
+import dynamic from 'next/dynamic';
 
 // components
 import Layout from 'layout/layout/layout-app';
@@ -23,7 +24,7 @@ import ExploreMyData from 'layout/explore/explore-my-data';
 import { Media } from 'lib/media';
 
 // constants
-import { EXPLORE_SECTIONS, EXPLORE_SUBSECTIONS } from './constants';
+import { EXPLORE_SECTIONS, EXPLORE_SUBSECTIONS, DATASETS_WITH_SCHEMA } from './constants';
 
 const Explore = (props) => {
   const {
@@ -82,6 +83,14 @@ const Explore = (props) => {
     </>
   );
 
+  const DatasetSchemaScript = useMemo(() => {
+    if (!datasetData?.slug || !DATASETS_WITH_SCHEMA.includes(datasetData?.slug)) {
+      return null;
+    }
+
+    return dynamic(() => import(`scripts/schemas/${datasetData.slug.toLowerCase()}`));
+  }, [datasetData]);
+
   const metadata = dataset && dataset.metadata && dataset.metadata[0];
   const infoObj = metadata && metadata.info;
   const titleSt = selected ? infoObj && infoObj.name : '';
@@ -91,6 +100,7 @@ const Explore = (props) => {
 
   return (
     <Layout title={titleSt} description={descriptionSt} className="-fullscreen" isFullScreen>
+      {DatasetSchemaScript && <DatasetSchemaScript />}
       <Head>
         {/* unpublished datasets are not indexed by search engines but still accessible in the application */}
         {datasetData && !datasetData?.published && <meta name="robots" content="noindex, follow" />}